Test E2 items in E2 only. Remove eye of demon (broken item).
