+++ res/core/en/strings.xml
@@ -444,12 +444,6 @@
   <string name="ao_chastity_p">
     <text locale="en">amulets of chastity</text>
   </string>
-  <string name="aoc">
-    <text locale="en">amulet of the kitten</text>
-  </string>
-  <string name="aoc_p">
-    <text locale="en">amulets of the kitten</text>
-  </string>
   <string name="aod">
     <text locale="en">amulet of darkness</text>
   </string>
