+++ scripts/tests/e2/init.lua
@@ -8,6 +8,7 @@ require 'tests.e2.destroy'
 require 'tests.e2.guard'
 require 'tests.e2.spells'
 require 'tests.e2.stealth'
+require 'tests.e2.items'
 require 'tests.items'
 require 'tests.orders'
 require 'tests.common'
